Fix upgrades for multinode galera
In an upgrade of multi-node galera cluster, the
galera_server_bootstrap_node will force a bootstrap even if the cluster
is running normally on the existing other nodes. This will fail on the
bootstrap_host and block the upgrade.

Lets test if the rest of the cluster is up by using a secondary node
that isn't the bootstrap node. If the rest of the cluster is available
we should skip the bootstrap tasks.

Change-Id: Ic4d69f0fa75c1eea81d10a76cca9a8d9c3822094
Closes-Bug: #1667103